+++ lib/delighted/operations/list.rb
@@ -1,0 +1,15 @@
+module Delighted
+  module Operations
+    module List
+      def self.included(klass)
+        klass.extend(ClassMethods)
+      end
+
+      module ClassMethods
+        def list(opts = {}, client = Delighted.shared_client)
+          ListResource.new(self, path, Utils.serialize_values(opts), client)
+        end
+      end
+    end
+  end
+end